Updates 10-18-2015
OPMODE5220:

Finished most of the move and rotate methods to the best of our current
abilities. Now there are methods to move with encoder or time, and to
rotate with gyro, encoder, or time. The encoder and gyro reading methods
are still just nearly empty placeholders that need to be replaced as
soon as we have the encoder and gyro sensor working.

Added methods to convert from inches to (drivetrain) encoder counts. All
the calculations are based on the WHEEL_DIAMETER, GEAR_RATIO, and
ENCODER_COUNTS_PER_ROTATION constants. The conversion methods are
currently applied in the encoder-controlled move and rotate methods.

Added a “global” variable to indicate the phase of the program, which is
automatically updated in the setup, initialization, and waitForStart
(overridden to do so) methods, in addition to the main runOpMode method
body.

Eliminated the “special” setup and initialization methods in favor of
overriding setup and initialization in subclasses. It is important to
remember to call super.setup() and/or super.init() in the setup and
initialization methods in overriding subclasses before doing anything
else .

AUTONOMOUS:

Added ConfigLoop. This loop allows the user to configure program
settings using two touch sensors. Currently, the main loop and its logic
are fully fleshed out, but the touch sensor reading method remains a
placeholder.  Essentially, one touch sensor increments a setting, the
other decrements a setting, and if both sensors are pressed, the program
will move on to the next configuration setting. The current settings and
which setting is being modified should be shown in driver station
telemetry, but the telemetry feedback has NOT been written yet. The loop
will exit once the main program is started.

TELEOP:

Updated Tele-Op exclusive initialization for the aforementioned removal
of the specialInitialization method.

OTHER:

Saved a separate “original” version of FtcRobotControllerActivity and
created a new copy, in case we want to make potentially risky changes to
the new one (and assuming that we are allowed to make those changes).

A few other minor improvements not worthy of specific documentation.