First Working Program For The 3-Axis Debris Handling Mechanism and Big Collector
ROBOTCONFIG:

Ran phone configuration for the new motor setup (with the big sweeper
and the 3-axis arm) and changed robot_config.xml on the computer to
match the xml file from the phone.

OPMODE5220:

Added the new motor controller and the new servo controller. The motor
controller is on port 2 of the power module, and the servo controller is
on port 3. On the motor controller, port 1 is the motor that
horizontally rotates the arm, and port 2 is the motor that  rotates the
sweeper. On the servo controller, port 1 is the Quarter-scale (more
powerful) servo that rotates the last portion of the arm upwards, and
port 2 is the (normal) servo that rotates the block/ball holder only.

Changed the names of all the motors, servos, and controllers. All the
motor and servo controllers are now set to their default names in phone
configuration to save time in phone configuration. Also, the motor and
servo names in phone configuration (and robot_config.xml)are now much
shorter (to save time in phone configuration), and they reflect their
respective purposes on  the robot.

TELEOP:

Added control of the arm and sweeper. Currently, it is functional but
quite troublesome for the driver to use. If we stick with the 3 axis
design, control of the arm definitely needs to be made easier for the
driver.

Commented out the condition that makes the TeleOp program end after 2
minutes. We're not going to need to use the time limit until we develop
a robot good enough that we want to field test its speed, which will
probably take a while.