Updates 10-22-2015
Added motor syncing for moving forwards and backward,  which can be done
with the encoders (all 4 of them) or the gyro sensor. This should help
us to move straighter. Currently, the gyro reading method is still a
placeholder method that always returns a fixed value.

This new motor syncing has been integrated into a new, universally
applicable move method than can be used for any forwards or backwards
motion at any power, with or without syncing.