Updates 10-26-2015
ROBOT_CONFIG:

Removed the old DC motor for swiveling the arm, and replaced it with the
new servo for swiveling the arm.

OPMODE5220:

Added DebuggerDisplayLoop to display and continuously update important
data in telemetry lines 1-8.

Replaced all setChannelMode (RUN_WITHOUT_ENCODER) usages with
setChannelMode (RUN_USING_ENCODERS).  Now the encoders sort of work

Changed Stopwatch time() method to return elapsed time in milliseconds,
and added timeSeconds() method, which gives the time in seconds.

Added something of a functional yet far from complete placeholder test
(using only left back encoder) in driveEncodersHaveReached.

Added moveSimple method, which just moves a given encoder count without
distance conversion or the option of syncing. Once the integrated move()
method is working, this method can be eliminated.

Added placeholder method to automatically move the arm to the collection
position or a reasonable dispensing position.

Still need to get rid of tiltServo for the new tube design.

TELEOP:

Added slow movement option: the robot’s maximum drive power is equal to
the constant SLOW_POWER (currently 0.15) while the left stick button is
pressed.

Changed how the driver controls the arm. The new way is far more
programmatically complicated, but MUCH easier for the driver. Now (as
opposed to the impossibly difficult old way of having to hold the
joystick in exactly the right position to stay still) the driver can
simply control the arm with the top hat: left and right control
swiveling, and up and down control up/down rotation. If nothing is
pressed on the game pad top hat the arm will stay still. If the arm is
somehow pushed out of position without gear slipping, it will
automatically return to where it was supposed to be (if that is
physically possible). Also, if the arm is having a hard time reaching
its target position (perhaps due to gravity), it will increase its power
to reach the target position.

All the constants for the new arm control scheme (both swiveling and
up/down) are conveniently set near the beginning of the program.

Added setChannelMode(RUN_USING_ENCODERS) for all drivetrain motors
(redundant from OpMode5220)

Added the use of DebuggerDisplayLoop()

AUTONOMOUS:

Removed system.exit everywhere, since it was messing up the program,
making only one run of autonomous to work per opening of the Robot
Controller App.

Added use of DebuggerDisplayLoop

Added some simple tests in the test() method for moveSimple and move. As
of this commit, moveSimple works okay (so the encoders do work), but
move() does not.

OTHER:

Removed Old_5220_OpMode from the OpMode Register.

Created LoopTest5220, a clone of the old TeleOp (approximately what it
was upon the last GitHub commit) that uses OpMode and loop() instead of
LinearOpMode, and added it to the OpMode register. This was done to try
to see if LinearOpMode was messing up the encoders, but it turns out
that the encoders were just plugged in the wrong way. So LoopTest5220
can be deleted, and removed from the Op Mode Register. As of this commit
however, that has not yet been done.

Probably made a few minor changes not worthy of documentation, in
addition to all the stuff listed above.