Updates 11-9-2015
AUTONOMOUS:

Added getRescueBeaconColor method to try judge whether the rescue beacon
color is red, blue, or indeterminate. Currently, it is only based on
what seems like it should work, rather than testing. Quite a bit of
testing will probably be neccessary to make this method the best that it
can be.

TELEOP:

Added temporary fix for the swivel inertia problem (halving the final
value). Not sure whether to keep it or not.

OPMODE5220:

Added colorSensor and its initialization.

Added color sensor readings to DebuggerDisplayLoop

OTHER:

Fixed DriveControl_5220 (it had a while condition that would not work
properly before).

Added the Device Interface Module and the Color Sensor to
robot_config.xml (of course)

Copy pasted ColorSensorDriver into our own version called
ColorSensorDriver_5220, modified it so that it would actually work with
our robot, and added it to the Op Mode Register.

Added Config_3 and Config_4 text files to main directory, so
robot_config.xml can be switched between them easily if neccessary.
Ideally all the configurations we might actually use should be stored
similarly to Config_3 and Config_4 so we can quickly copy past them into
robot_config.xml without having to copy over the config file from the
phone.