Updates 11-14-2015
Added new system for keeping encoder initial values and performing
operations based on them, since there seems to be no way to reset the
actual encoders.

Fixed problem with GEAR_RATIO constant being zero instead of 2.0/3.0

Added use of debug log.

Added test for any potential changes we might make to the UI in
OpMode5220. this should be moved to ftcrobotcontrolleractivity though.

Added placeholder method to load any modifications to the GUI we might
want to make in the near future in FtcRobotControllerActivity.

Made various temporary changes in the autonomous program for testing.