Updates 11-22-2015
Added Gamepad 2 controls for moving the arm to set positions and
extending the hooks. More gamepad 2 controls to be added soon.

Fixed Swivel_360. Before, the value of Swivel_360 was actually the value
for turning 180 degrees.

Added use of Gyro Sensor. The gyro rotation methods still need to be
fixed. It may be necessary to put the gyro in a separate Thread that
continuously feed the ABSOLUTE (not mod 360) gyro rotation value into a
global variable. Alternatively, we could do the integration of the Z
axis ourselves, port from RobotC, instead of using the built in one and
run that in the separate thread.

Fixed problem with feeding negative degree values to rotate method

Very minor value changes to Autonomous.

Updated Robot_Config for Gyro sensor addition.