Updates 11-26 2015
Changed HOOK_TILT_INCREMENT from 0.025 to 0.072.

Added touch sensors, and they have been tested to work okay with the
autonomous ConfigLoop.

Added use of the climber flinging servo.

Added an autnonomous program, currently in the test() method, to go to
the rescue beacon area and try to fling the climbers in.

Updated robot_config on the phone and the computer for the addition of
the touch sensors and the new climber flinger.