Updates 11-29-2015
Deleted unnecessary PDFs from the repository.

Updated Robot config on both phone and computer for addition of the
climber trigger servos.

Added some new controls for driver 2, as documented in Quick_Guide.txt

Modified climber fling to go slower using a timer in order to avoid
throwing the climbers too far.

Fine tuned all autonomous movement values to 85-90% reliability.
Currently, the autonomous program deposits both climbers in the goal and
drives up the ramp. There are still many waits in the program that can
be eliminated soon.

Improved Config_Loop a bit in autonomous.

Added use of error method (instead of msg) for Debug Log, since it seems
like the only difference is that error messages are in red, and thus
more noticeable.

Changed GEAR_RATIO from 2/3 to 1/3 and adjusted movement values
accordingly.

Moved current autonomous code from test() to autonomous(). We still need
to add the code for when we are RED (instead of BLUE).