Updates 1-9-2016
OpMode_5220: Changed CLIMBER_FLING_TIME from 3.0 to 1.0 (seconds), added
ftcRCA (the FTCRobotControllerActivity) as a protected variable, and
added a currently unused method to restart the robot using ftcRCA.

Autonomous: Fine-tuned the values for BLUE, copy pasted those values to
RED, reversed them, and fine-tuned that.

Drive Control: Changed forwards/backwards power from 0.5 to 0.7, and
changed turn power from 0.5 to 1.0.

FTCRobotControllerActivity: Added static variable ftcRCA that is set in
the constructor to the newest instance of FTCRobotControllerActivity
when it (the Activity) is created.