Final Nevada Championship Software
This was the code that we went into the elimination matches at the Las
Vegas tournament with. We did not make any more changes after the start
of the elimination matches.

We need to add the waitOneFullHardwareCycle thing everywhere. Before
doing so though, we should condense into a method with a shorter name
that has the try catch included.