Updates 2-11-2016
Added new programFinished global variable, which runConditions depends
on. This way, the program can effectively be ended neatly from anywhere
inside it.

Added detection of stalling while following the line (caused by the
occasional buildup of three blocks in front of the plow/wall). If we
have been trying to follow the line forwards but the touch sensor has
not yet been pressed yet at the 17 second mark, we simply extend the
linear slides forward, score the climbers, and back out.