Updates 2-11-2016 2
Changed default color back to red.

Added starting positions to ConfigLoop: currently there are two options
(starting near the ramp and starting near the corner), but more can be
added in the unlikely case that such is required.

Added the multiple starting positions to the autonomous program. The
robot can start in multiple locations, but once the robot reaches the
line,  the programs for different starting positions merge back into
one.

Note: the values that we have fine-tuned up to this time are for
starting near the ramp. Those fine-tuned values are preserved in the
case in which we start near the ramp, while the values for starting near
the corner are completely untested as of now.