Updates 2-21-2016
Change a lot of things in autonomous to try to get red autonomous to
work. Red corner autonomous simply cannot work consistently with turning
inaccuracy, as of yet. Blue autonomous works ok, but still is not as
consistent and precise as it should be.

Reversed sweeper motors (for some reason this was necessary)

Changed TeleOp sweeper power from 1.0 to 0.8