Updates 3-20-2016
Changed gear ratio for the new drivetrain. The drivetrain is supposed to
have changed from 1/2 gear ratio to a 2/3 gear ratio, but it actually
seems to be a 4/5 gear ratio now.

Added global variable to turn of debugger display loop if necessary.

Added method to get the average of the left or right encoders.

Restructured driveEncodersHaveReached to use new sided average methods

Rewrote turnEncodersHaveReached to work like the new
driveEncodersHaveReached, using the side averages.

Experimentally changed moveSmooth smoothDistance to be a third of the
given distance, instead of a constant.

Changed moveSmooth max power to 0.75

Wrote new Red Ramp autonomous program for getting to the beacon area. It
*almost* works.