Updates 4-8-2016
Added music playing (for fun).

Added NavX Micro IMU initialization and telemetry display, and tested
that it works. Still have to add code in autonomous to use the sensor
values.

Restructured Debug Display Loop a bit to add the IMU readings, and place
the values in a more logical order.