Updates 4-17-2016
Modified everything involving IMU rotation a lot., to make it all work.

We now use a low-power correction loop after the initial high-power turn
in rotateIMU which ensures that the the method ends with the robot
within 0.4 degrees of its intended heading.

DEFAULT_TURN_POWER_HIGH is now 1.0 (before, it was 0.8).

Replaced gyro syncing in move method with IMU syncing.

Deleted the outdated, commented out sections of the test method.

Added test method for rotateIMU.